Factorize code with UglifyJSPlugin
