+++ lib/ModuleFilenameHelpers.js
@@ -41,6 +41,11 @@ function getHash(str) {
 	return hash.digest("hex").substr(0, 4);
 }
 
+function asRegExp(test) {
+	if(typeof test == "string") test = new RegExp("^"+test.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&"));
+	return test;
+}
+
 ModuleFilenameHelpers.createFilename = function createFilename(module, moduleFilenameTemplate, requestShortener) {
 	if(typeof module === "string") {
 		var shortIdentifier = requestShortener.shorten(module);
@@ -113,4 +118,27 @@ ModuleFilenameHelpers.replaceDuplicates = function replaceDuplicates(array, fn, 
 			return fn(item, i, posMap[item]++);
 		} else return item;
 	});
+};
+
+
+ModuleFilenameHelpers.matchPart = function matchPart(str, test) {
+	if(!test) return true;
+	test = asRegExp(test);
+	if(Array.isArray(test)) {
+		return test.map(asRegExp).filter(function(regExp) {
+			return regExp.test(str);
+		}).length > 0;
+	} else {
+		return test.test(str);
+	}
+};
+
+ModuleFilenameHelpers.matchObject = function matchObject(obj, str) {
+	if(obj.test)
+		if(!ModuleFilenameHelpers.matchPart(str, obj.test)) return false;
+	if(obj.include)
+		if(!ModuleFilenameHelpers.matchPart(str, obj.include)) return false;
+	if(obj.exclude)
+		if(ModuleFilenameHelpers.matchPart(str, obj.exclude)) return false;
+	return true;
 };
