+++ lib/SourceMapDevToolPlugin.js
@@ -34,8 +34,8 @@ SourceMapDevToolPlugin.prototype.apply = function(compiler) {
 	var fallbackModuleFilenameTemplate = this.fallbackModuleFilenameTemplate;
 	var requestShortener = new RequestShortener(compiler.context);
 	var cheapMode = this.cheapMode;
-	var that = this;
 	var options = this.options;
+	options.test = options.test || /\.js($|\?)/i;
 	compiler.plugin("compilation", function(compilation) {
 		if(cheapMode) {
 			compilation.moduleTemplate.plugin("module", function(source, module) {
@@ -52,7 +52,7 @@ SourceMapDevToolPlugin.prototype.apply = function(compiler) {
 			var allModuleFilenames = [];
 			var tasks = [];
 			chunks.forEach(function(chunk) {
-				chunk.files.filter(that.matchObject.bind(that, options)).map(function(file) {
+				chunk.files.filter(ModuleFilenameHelpers.matchObject.bind(undefined, options)).map(function(file) {
 					var asset = this.assets[file];
 					if(asset.__SourceMapDevTool_Data) {
 						var data = asset.__SourceMapDevTool_Data;
@@ -162,28 +162,6 @@ SourceMapDevToolPlugin.prototype.apply = function(compiler) {
 			}, this);
 		});
 	});
-};
-
-function asRegExp(test) {
-	if(typeof test == "string") test = new RegExp("^"+test.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&"));
-	return test;
-}
-SourceMapDevToolPlugin.prototype.matchPart = function matchPart(str, test) {
-	if(!test) return true;
-	test = asRegExp(test);
-	if(Array.isArray(test)) {
-		return test.map(asRegExp).filter(function(regExp) {
-			return regExp.test(str);
-		}).length > 0;
-	} else {
-		return test.test(str);
-	}
-};
-
-SourceMapDevToolPlugin.prototype.matchObject = function matchObject(obj, str) {
-	if(obj.exclude)
-		if(this.matchPart(str, obj.exclude)) return false;
-	return true;
 };
 
 function basename(name) {
