+++ lib/optimize/UglifyJsPlugin.js
@@ -6,6 +6,7 @@ var SourceMapConsumer = require("webpack-core/lib/source-map").SourceMapConsumer
 var SourceMapSource = require("webpack-core/lib/SourceMapSource");
 var RawSource = require("webpack-core/lib/RawSource");
 var RequestShortener = require("../RequestShortener");
+var ModuleFilenameHelpers = require("../ModuleFilenameHelpers");
 var uglify = require("uglify-js");
 
 function UglifyJsPlugin(options) {
@@ -19,7 +20,6 @@ module.exports = UglifyJsPlugin;
 
 UglifyJsPlugin.prototype.apply = function(compiler) {
 	var options = this.options;
-	var that = this;
 	options.test = options.test || /\.js($|\?)/i;
 
 	var requestShortener = new RequestShortener(compiler.context);
@@ -40,7 +40,7 @@ UglifyJsPlugin.prototype.apply = function(compiler) {
 			compilation.additionalChunkAssets.forEach(function(file) {
 				files.push(file);
 			});
-			files = files.filter(that.matchObject.bind(that, options));
+			files = files.filter(ModuleFilenameHelpers.matchObject.bind(undefined, options));
 			files.forEach(function(file) {
 				var oldWarnFunction = uglify.AST_Node.warn_function;
 				var warnings = [];
@@ -132,31 +132,4 @@ UglifyJsPlugin.prototype.apply = function(compiler) {
 			context.minimize = true;
 		});
 	});
-};
-
-function asRegExp(test) {
-	if(typeof test == "string") test = new RegExp("^"+test.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&"));
-	return test;
-}
-
-UglifyJsPlugin.prototype.matchPart = function matchPart(str, test) {
-	if(!test) return true;
-	test = asRegExp(test);
-	if(Array.isArray(test)) {
-		return test.map(asRegExp).filter(function(regExp) {
-			return regExp.test(str);
-		}).length > 0;
-	} else {
-		return test.test(str);
-	}
-};
-
-UglifyJsPlugin.prototype.matchObject = function matchObject(obj, str) {
-	if(obj.test)
-		if(!this.matchPart(str, obj.test)) return false;
-	if(obj.include)
-		if(!this.matchPart(str, obj.include)) return false;
-	if(obj.exclude)
-		if(this.matchPart(str, obj.exclude)) return false;
-	return true;
 };
